+++ packages/icons/eslint.config.mjs
@@ -1,0 +1,9 @@
+import configs from '@status-im/eslint-config'
+
+/** @type {import('eslint').Linter.Config[]} */
+export default [
+  ...configs,
+  {
+    ignores: ['dist', 'index.js', 'index.es.js'],
+  },
+]