+++ packages/status-js/src/utils/decompress-public-key.ts
@@ -4,7 +4,7 @@ export function decompressPublicKey(publicKey: string): string {
   try {
     const pk = publicKey.replace(/^0[xX]/, '') // ensures hexadecimal digits without "base prefix"
     return secp.Point.fromHex(pk).toHex()
-  } catch (error) {
+  } catch {
     throw new Error('Invalid public key')
   }
 }