+++ packages/status-js/src/utils/payload-to-id.ts
@@ -3,7 +3,7 @@ import { bytesToHex, concatBytes } from 'ethereum-cryptography/utils'
 
 export function payloadToId(
   payload: Uint8Array,
-  publicKey: Uint8Array
+  publicKey: Uint8Array,
 ): string {
   const hash = keccak256(concatBytes(publicKey, payload)) // order matters
   const hex = bytesToHex(hash)