+++ packages/status-js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @status-im/js
 
+## 1.0.2
+
+### Patch Changes
+
+- 3c71b8c: update ESLint
+
+## 1.0.1
+
+### Patch Changes
+
+- 2fad122: use pnpm and update node
+
 ## 1.0.0
 
 ### Major Changes